week2 - 변수 이름 변경
